move document picker
