+++ src/backoffice/documents/documents/components/index.ts
@@ -1,0 +1,1 @@
+export * from './input-document-picker/input-document-picker.element';
