+++ src/backoffice/documents/documents/components/input-document-picker/input-document-picker.stories.ts
