+++ src/backoffice/documents/documents/property-editors/document-picker/manifests.ts
