+++ src/backoffice/documents/documents/property-editors/document-picker/property-editor-ui-document-picker.element.ts
@@ -1,9 +1,8 @@
 import { html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
-import type { UmbInputDocumentPickerElement } from '../../../components/input-document-picker/input-document-picker.element';
+import type { UmbInputDocumentPickerElement } from '../../components/input-document-picker/input-document-picker.element';
 import { UmbPropertyEditorExtensionElement } from '@umbraco-cms/backoffice/extensions-registry';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
-import '../../../components/input-document-picker/input-document-picker.element';
 import type { DataTypePropertyPresentationModel } from '@umbraco-cms/backoffice/backend-api';
 
 @customElement('umb-property-editor-ui-document-picker')
