+++ src/backoffice/documents/documents/property-editors/document-picker/property-editor-ui-document-picker.stories.ts
