+++ src/backoffice/shared/property-editors/uis/manifests.ts
@@ -1,5 +1,4 @@
 import { manifest as colorPicker } from './color-picker/manifests';
-import { manifest as documentPicker } from './document-picker/manifests';
 import { manifest as datePicker } from './date-picker/manifests';
 import { manifest as eyeDropper } from './eye-dropper/manifests';
 import { manifest as multiUrlPicker } from './multi-url-picker/manifests';
@@ -36,7 +35,6 @@ import type { ManifestPropertyEditorUI } from '@umbraco-cms/backoffice/extension
 
 export const manifests: Array<ManifestPropertyEditorUI> = [
 	colorPicker,
-	documentPicker,
 	datePicker,
 	eyeDropper,
 	multiUrlPicker,
