+++ src/backoffice/users/users/workspace/user-workspace.element.ts
@@ -7,7 +7,7 @@ import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import type { UmbRoute } from '@umbraco-cms/backoffice/router';
 
 import '../../../shared/components/input-user-group/input-user-group.element';
-import '../../../shared/property-editors/uis/document-picker/property-editor-ui-document-picker.element';
+import '../../../documents/documents/property-editors/document-picker/property-editor-ui-document-picker.element';
 import '../../../shared/components/workspace/workspace-editor/workspace-editor.element';
 
 @customElement('umb-user-workspace')
